reafctor: extraido calculo de distância entre pontos para classe de utilidade
